Found even more casts. They sure aren't all of them, but I hope to have caught every pointer C-style cast because they can be very dangerous. Note: I didn't do the pointer casts in the network library because that would have taken way too long.

+++ src/core/IRC.cc
@@ -51,7 +51,7 @@ namespace orxonox
 
     void IRC::initialize()
     {
-        unsigned int threadID = (unsigned int)IRC_TCL_THREADID;
+        unsigned int threadID = IRC_TCL_THREADID;
         TclThreadManager::createID(threadID);
         this->bundle_ = TclThreadManager::getInstance().getInterpreterBundle(threadID);
